pyproject.toml
[project]
name = "coach"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
]
